util: extract ghost file copy logic
Ghost files logic is extracted from files-reg.c to util.c.  This
improves code organization, making writing new features easier, like
incremental checkpoints.
